Refactoring
- declared login layout inside packages > ui
