--- apps/core/src/pages/onboarding/components/LoginLayout.tsx
@@ -1,125 +1,0 @@
-import {
-  Paper,
-  createStyles,
-  TextInput,
-  PasswordInput,
-  Button,
-  Title,
-  Text,
-  Anchor,
-  Divider,
-  Flex
-} from '@mantine/core';
-import FormHeader from './FormHeader';
-import FormHeading from './FormHeading';
-import { Outlet } from 'react-router-dom';
-
-const useStyles = createStyles((theme) => ({
-    loginPage: {
-        height: '100vh',
-        background: '#FCF7F1',
-    },
-    
-    wrapper: {
-        width: '100%',
-        height: '100%',
-        minHeight: '100vh',
-    },
-
-    formContainer: {
-        width: '60%',
-        height: '100%',
-        positon: 'relative',
-        padding: '17px',
-        [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-            width: '100%',
-        },
-    },
-
-    formHeader: {
-        position: 'fixed',
-        left: '0',
-        right: '0',
-        top: '0',
-        height: '75px',
-        borderBottom: '1px solid #F9DBB6',
-        display: 'flex',
-        justifyContent: 'start',
-        alignItems: 'center',
-        paddingLeft: '110px',
-        [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-            paddingLeft: '17px',
-        },
-    },
-
-    formLogo: {
-        display: 'flex',
-        gap: '10px',
-        justifyContent: 'start',
-        alignItems: 'center',
-        
-    },
-
-    form: {
-        // borderRight: `1px solid ${
-        // theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.colors.gray[3]
-        // }`,
-        maxWidth: '480px',
-        width: '100%',
-        background: 'transparent',
-        [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-        maxWidth: '480px',
-        },
-    },
-
-    formImage: {
-        width: '40%',
-        height: '100%',
-        background: '#14466F',
-        color: 'white',
-        zIndex: 99,
-        [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-            display: 'none',
-        },
-    },
-
-    logo: {
-        color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-        width: 120,
-        display: 'block',
-        marginLeft: 'auto',
-        marginRight: 'auto',
-    },
-
-    label: {
-        fontSize: '20px',
-        fontWeight: 700,
-        color: '#373B3D'
-    }
-}));
-
-const LoginLayout = () => {
-  const { classes } = useStyles();
-  return (
-    <div className={classes.loginPage}>
-        <Flex
-            className={classes.wrapper}
-            justify="between"
-            align="center"
-        >
-            <Flex 
-                className={classes.formContainer}
-                justify="center"
-                align="center"
-            >
-                <FormHeader />
-                <Outlet />  
-            </Flex>
-            <Paper className={classes.formImage}>
-            </Paper>
-        </Flex>
-    </div>
-  );
-}
-
-export default LoginLayout
